requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
sqlalchemy
asyncpg
aioredis
slowapi
fastapi-cache2
redis